All pages done. Made Your Orders page, Invoice page to save an order and made corrections and added Modals

+++ src/main/java/com/capgemini/go/model/OrderModel.java
@@ -52,7 +52,7 @@ public class OrderModel {
 //	@JsonIgnore
 	private List<ProductModel> products;
 	
-	private Long totalQuantity;
+	private Integer totalQuantity;
 	private Double totalPrice;
 	private LocalDate orderDeliveryDate;
 	private LocalDate orderDispatchDate;
